Fix issue with failing ChessBoard tests...
The root cause stemmed from the probability that a piece
would be replaced with the exact same piece. When this happens,
the board is not actually different, and test would fail improerly.

Since there are 6 pieces * 2 colors = 12 pieces, the probabiliy
of an individual piece being replaced with itself is 1 in 12.
This is about 9%. However, there are 14 different pieces that
each have this probability. The total probability compounds
up to ~66%!

Because of this overlooked error, the tests would fail improperly
most of the time.

This change does not call resetBoard() before creating each of the boards.
Therefore, they are all blank, and adding a random piece in a different
position each time always creates a definitely different board.