Add ChessGameTests...
These apply the EqualsTestingUtility class to ChessGame.
Although the equals() and hashCode() methods aren't
_strictly_ necessary for future phases, this is the
easiest, most reasonable and consistent way to ensure
that no fields are defined as 'static'.

Additionally, it doesn't hurt to have students
remember how to generate the two methods again.

Adding another test here isn't anymore of a burden
on students getting started on Phase 1 because we
already instruct them to copy over the entire
'game/' directory in the instructions.